Fix & refactoring routes Add readme
