+++ README.md
@@ -1,0 +1,27 @@
+# Mesto_NodeJS
+Версия 0.0.1
+
+## О проекте:
+разработка сервера на NodeJS для учебного проекта Mesto
+
+
+## Основной функционал:
+- Обработка GET запросов
+- раздача статических файллов
+
+## Стэк технологий:
+NodeJS, express
+
+## Пакеты которые используются в сборках:
+
+  - "eslint": "^7.8.1",
+  - "eslint-config-airbnb-base": "^14.2.0",
+  - "eslint-plugin-import": "^2.22.0",
+  - "nodemon": "^2.0.4"
+  - "express": "^4.17.1"
+
+## Инструкции по запуску:
+- Скачать или склонировать репозиторий
+- Установить зависимости при помощи npm - `npm i`
+- Команда npm run start запускает сервер на localhost:3000
+- Команда npm run dev запускает сервер на localhost:3000 с хот релоудом
